requirements.txt
langgraph>=1.0.6
langgraph-checkpoint>=4.0.0
langgraph-checkpoint-postgres>=2.0.0
langchain-core>=1.2.7
psycopg[binary,pool]>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
